Move users models to the users namespace
AB#15737
