+++ src/NHSD.GPIT.BuyingCatalogue.EntityFramework/Users/Models/AspNetRole.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.AspNetCore.Identity;
 
-namespace NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue
+namespace NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models
 {
     [Table("AspNetRoles")]
     public class AspNetRole : IdentityRole<string>
