+++ src/NHSD.GPIT.BuyingCatalogue.EntityFramework/Users/Models/AspNetRoleClaim.cs
@@ -2,12 +2,12 @@
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Identity;
 
-namespace NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue
+namespace NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models
 {
     [ExcludeFromCodeCoverage(Justification = "Class currently only contains automatic properties")]
-    [Table("AspNetUserLogins")]
-    public partial class AspNetUserLogin : IdentityUserLogin<string>
+    [Table("AspNetRoleClaims")]
+    public partial class AspNetRoleClaim : IdentityRoleClaim<string>
     {
-        public virtual AspNetUser User { get; set; }
+        public virtual AspNetRole Role { get; set; }
     }
 }
