+++ src/NHSD.GPIT.BuyingCatalogue.EntityFramework/Users/Models/AspNetUserLogin.cs
@@ -2,14 +2,12 @@
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.AspNetCore.Identity;
 
-namespace NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue
+namespace NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models
 {
     [ExcludeFromCodeCoverage(Justification = "Class currently only contains automatic properties")]
-    [Table("AspNetUserRoles")]
-    public partial class AspNetUserRole : IdentityUserRole<string>
+    [Table("AspNetUserLogins")]
+    public partial class AspNetUserLogin : IdentityUserLogin<string>
     {
-        public virtual AspNetRole Role { get; set; }
-
         public virtual AspNetUser User { get; set; }
     }
 }
