+++ src/NHSD.GPIT.BuyingCatalogue.Framework/Extensions/AspNetUserExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 
 namespace NHSD.GPIT.BuyingCatalogue.Framework.Extensions
 {
