+++ src/NHSD.GPIT.BuyingCatalogue.ServiceContracts/CreateBuyer/IAspNetUserValidator.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 using NHSD.GPIT.BuyingCatalogue.ServiceContracts.Results;
 
 namespace NHSD.GPIT.BuyingCatalogue.ServiceContracts.CreateBuyer
