+++ src/NHSD.GPIT.BuyingCatalogue.ServiceContracts/Identity/PasswordResetToken.cs
@@ -1,4 +1,5 @@
 ﻿using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 
 namespace NHSD.GPIT.BuyingCatalogue.ServiceContracts.Identity
 {
