+++ src/NHSD.GPIT.BuyingCatalogue.Services/CreateBuyer/AspNetUserErrors.cs
@@ -1,4 +1,5 @@
 ﻿using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 using NHSD.GPIT.BuyingCatalogue.ServiceContracts.Errors;
 
 namespace NHSD.GPIT.BuyingCatalogue.Services.CreateBuyer
