+++ tests/NHSD.GPIT.BuyingCatalogue.E2E.PublicBrowseTests/Utils/RandomData/GenerateUser.cs
@@ -3,6 +3,7 @@ using Bogus;
 using Microsoft.AspNetCore.Identity;
 using NHSD.GPIT.BuyingCatalogue.E2ETests.Utils.TestModels;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 
 namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Utils.RandomData
 {
