+++ tests/NHSD.GPIT.BuyingCatalogue.Framework.UnitTests/Extensions/AspNetUserExtensionsTests.cs
@@ -1,5 +1,6 @@
 ﻿using FluentAssertions;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 using NHSD.GPIT.BuyingCatalogue.Framework.Extensions;
 using Xunit;
