+++ tests/NHSD.GPIT.BuyingCatalogue.Services.UnitTests/CreateBuyer/AspNetUserValidatorTests.cs
@@ -7,6 +7,7 @@ using FluentAssertions;
 using Moq;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 using NHSD.GPIT.BuyingCatalogue.ServiceContracts.Errors;
 using NHSD.GPIT.BuyingCatalogue.ServiceContracts.Results;
 using NHSD.GPIT.BuyingCatalogue.Services.CreateBuyer;
