+++ tests/NHSD.GPIT.BuyingCatalogue.Services.UnitTests/Users/UsersServiceTests.cs
@@ -5,6 +5,7 @@ using FluentAssertions;
 using Moq;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 using NHSD.GPIT.BuyingCatalogue.Services.Users;
 using NHSD.GPIT.BuyingCatalogue.Test.Framework.TestData;
 using Xunit;
