+++ tests/NHSD.GPIT.BuyingCatalogue.Test.Framework/Comparers/AspNetUserEditableInformationComparer.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Models.GPITBuyingCatalogue;
+using NHSD.GPIT.BuyingCatalogue.EntityFramework.Users.Models;
 
 namespace NHSD.GPIT.BuyingCatalogue.Test.Framework.Comparers
 {
